refactor: move flaeg only tests.
